+++ crates/rattler_networking/src/lib.rs
@@ -2,6 +2,7 @@
 
 //! Networking utilities for Rattler, specifically authenticating requests
 pub use authentication_middleware::AuthenticationMiddleware;
+pub use gcs_middleware::GCSMiddleware;
 pub use authentication_storage::{authentication::Authentication, storage::AuthenticationStorage};
 pub use mirror_middleware::MirrorMiddleware;
 pub use oci_middleware::OciMiddleware;
@@ -10,6 +11,7 @@ pub mod authentication_middleware;
 pub mod authentication_storage;
 pub mod mirror_middleware;
 pub mod oci_middleware;
+pub mod gcs_middleware;
 pub mod retry_policies;
 
 mod redaction;
